remove rpc specs
