+++ packages/api/src/executor/index.ts
@@ -1,5 +1,4 @@
 export * from './Executor.js';
-export * from './RpcExecutor.js';
 export * from './JsonRpcExecutor.js';
 export * from './ConstantExecutor.js';
 export * from './StorageQueryExecutor.js';
