+++ packages/codegen/src/generator/QueryGen.ts
@@ -1,6 +1,6 @@
 import { StorageEntryLatest } from '@dedot/codecs';
 import { normalizeName, stringCamelCase } from '@dedot/utils';
-import { ApiGen } from '../generator/index.js';
+import { ApiGen } from './ApiGen.js';
 import { beautifySourceCode, commentBlock, compileTemplate } from './utils.js';
 
 export class QueryGen extends ApiGen {
