+++ packages/codegen/src/generator/index.ts
@@ -2,7 +2,6 @@ export * from './TypesGen.js';
 export * from './ApiGen.js';
 export * from './ConstsGen.js';
 export * from './QueryGen.js';
-export * from './RpcGen.js';
 export * from './JsonRpcGen.js';
 export * from './IndexGen.js';
 export * from './ErrorsGen.js';
