+++ packages/codegen/src/templates/index.hbs
@@ -3,7 +3,6 @@
 import { GenericSubstrateApi } from '@dedot/types';
 import { ChainConsts } from './consts';
 import { ChainStorage } from './query';
-import { RpcCalls } from './rpc';
 import { ChainJsonRpcApis } from './json-rpc';
 import { ChainErrors } from './errors';
 import { ChainEvents } from './events';
@@ -13,7 +12,6 @@ import { ChainTx } from './tx';
 export * from './types';
 
 export interface {{{interfaceName}}}Api extends GenericSubstrateApi {
-  rpc: RpcCalls;
   jsonrpc: ChainJsonRpcApis;
   consts: ChainConsts;
   query: ChainStorage;
