--- packages/codegen/src/templates/rpc.hbs
@@ -1,7 +1,0 @@
-// Generated by @dedot/codegen
-
-{{{ importTypes }}}
-
-export interface RpcCalls extends GenericRpcCalls {
-  {{{rpcCallsOut}}}
-}
