+++ packages/specs/src/index.ts
@@ -1,3 +1,2 @@
-export * from './rpc/index.js';
 export * from './runtime/index.js';
 export * from './json-rpc/index.js';
