--- packages/specs/src/rpc/babe.ts
@@ -1,9 +1,0 @@
-import { RpcModuleSpec } from '@dedot/types';
-
-export const babe: RpcModuleSpec = {
-  epochAuthorship: {
-    docs: 'Returns data about which slots (primary or secondary) can be claimed in the current epoch with the keys in the keystore.',
-    params: [],
-    type: 'Record<string, EpochAuthorship>',
-  },
-};
