--- packages/specs/src/rpc/dev.ts
@@ -1,16 +1,0 @@
-import { RpcModuleSpec } from '@dedot/types';
-import { atBlockHashParam } from './shared.js';
-
-export const dev: RpcModuleSpec = {
-  getBlockStats: {
-    docs: [
-      'Reexecute the specified `block_hash` and gather statistics while doing so.\n',
-      '\n',
-      'This function requires the specified block and its parent to be available\n',
-      'at the queried node. If either the specified block or the parent is pruned,\n',
-      'this function will return `None`.',
-    ],
-    params: [atBlockHashParam],
-    type: 'Option<BlockStats>',
-  },
-};
