--- packages/specs/src/rpc/rpc.ts
@@ -1,9 +1,0 @@
-import { RpcModuleSpec } from '@dedot/types';
-
-export const rpc: RpcModuleSpec = {
-  methods: {
-    docs: 'Retrieves the list of RPC methods that are exposed by the node',
-    params: [],
-    type: 'RpcMethods',
-  },
-};
