--- packages/specs/src/rpc/shared.ts
@@ -1,7 +1,0 @@
-import { RpcParamSpec } from '@dedot/types';
-
-export const atBlockHashParam: RpcParamSpec = {
-  name: 'at',
-  type: 'BlockHash',
-  isOptional: true,
-};
